fix bugs detail product

+++ database/migrations/2020_06_19_185450_add_stars_rate_to_products.php
@@ -15,7 +15,7 @@ class AddStarsRateToProducts extends Migration
     {
         Schema::table('products', function (Blueprint $table) {
             //
-            $table->integer('stars_rate')->default(0);
+            $table->float('stars_rate')->default(0);
             $table->integer('count_rates')->default(0);
         });
     }
